requirements.txt
PyMySQL==0.10.0
kafka_python==2.0.1
kafka==1.3.5
PyYAML==5.3.1
